Add files via upload
One of the most difficult projects that I implemented was an empirical study conducted as part of an internship at Gazprom CPS LLC. The project was part of the digitalization program for welding production and executive documentation. 

The main objectives of the study included: 
1. Project pipeline development: organization of work processes, setting tasks and monitoring their implementation; 
2. Data processing: conducting primary analysis of data arrays on pipe welding operations; 
3. Model creation: development of a mock-up of a logistic regression model for classifying welding defects; 
4. Technical support: maintenance and monitoring of writing Python scripts; 
5. Analysis of results: performing exploratory data analysis (EDA), identifying the impact of technological parameters on the probability of failure, and evaluating the quality metrics of the machine learning model. 

The most difficult part of the project was the qualitative analysis of the data and the interpretation of the results obtained in a limited time. It was important not only to achieve high accuracy of the model, but also to make its conclusions understandable to the production staff. The results of the study allowed us to form practical recommendations for optimizing welding processes, which became an important contribution to the overall digitalization project.